# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Morex API",
        "description": "API pour l'application mobile Morex - Gestion des finances personnelles",
        "contact": {
            "name": "Morex Support",
            "email": "support@morex.app"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/categories": {
            "get": {
                "tags": [
                    "Categories"
                ],
                "summary": "Liste des catégories",
                "operationId": "b1b247b2b491b0491c3c728dc33f23f0",
                "responses": {
                    "200": {
                        "description": "Liste des catégories (personnelles + système)",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Category"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Non authentifié"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Categories"
                ],
                "summary": "Créer une catégorie",
                "operationId": "c86cd7a85ad06a4fe905bfab62146952",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "type"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Alimentation"
                                    },
                                    "type": {
                                        "type": "string",
                                        "enum": [
                                            "expense",
                                            "income"
                                        ]
                                    },
                                    "icon": {
                                        "type": "string",
                                        "example": "shopping-cart"
                                    },
                                    "color": {
                                        "type": "string",
                                        "example": "#10B981"
                                    },
                                    "parent_id": {
                                        "type": "string",
                                        "format": "uuid",
                                        "nullable": true
                                    },
                                    "budget_limit": {
                                        "description": "Budget mensuel en FCFA",
                                        "type": "integer",
                                        "example": 50000,
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Catégorie créée",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Category"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Erreur de validation"
                    },
                    "401": {
                        "description": "Non authentifié"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/categories/{id}": {
            "get": {
                "tags": [
                    "Categories"
                ],
                "summary": "Détail d'une catégorie",
                "operationId": "486849467f411ec5e4bdf111f50ede87",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Détail de la catégorie",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Category"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Catégorie non trouvée"
                    },
                    "401": {
                        "description": "Non authentifié"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Categories"
                ],
                "summary": "Modifier une catégorie",
                "operationId": "b648463d94e0caf7b8d842a1008789b2",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "type": {
                                        "type": "string",
                                        "enum": [
                                            "expense",
                                            "income"
                                        ]
                                    },
                                    "icon": {
                                        "type": "string"
                                    },
                                    "color": {
                                        "type": "string"
                                    },
                                    "parent_id": {
                                        "type": "string",
                                        "format": "uuid",
                                        "nullable": true
                                    },
                                    "budget_limit": {
                                        "type": "integer",
                                        "nullable": true
                                    },
                                    "order_index": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Catégorie modifiée",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Category"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Erreur de validation"
                    },
                    "403": {
                        "description": "Catégorie système non modifiable"
                    },
                    "404": {
                        "description": "Catégorie non trouvée"
                    },
                    "401": {
                        "description": "Non authentifié"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Categories"
                ],
                "summary": "Supprimer une catégorie",
                "operationId": "491dfe5b6b58b480181690ede6550e44",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Catégorie supprimée"
                    },
                    "403": {
                        "description": "Catégorie système non supprimable"
                    },
                    "404": {
                        "description": "Catégorie non trouvée"
                    },
                    "401": {
                        "description": "Non authentifié"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/categories/reorder": {
            "post": {
                "tags": [
                    "Categories"
                ],
                "summary": "Réordonner les catégories",
                "operationId": "2b55c7023140d0b064c7834663aa4414",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "categories"
                                ],
                                "properties": {
                                    "categories": {
                                        "type": "array",
                                        "items": {
                                            "properties": {
                                                "id": {
                                                    "type": "string",
                                                    "format": "uuid"
                                                },
                                                "order_index": {
                                                    "type": "integer"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Ordre mis à jour"
                    },
                    "401": {
                        "description": "Non authentifié"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Category": {
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "name": {
                        "type": "string",
                        "example": "Alimentation"
                    },
                    "type": {
                        "type": "string",
                        "enum": [
                            "expense",
                            "income"
                        ]
                    },
                    "color": {
                        "type": "string",
                        "example": "#10B981"
                    },
                    "icon": {
                        "type": "string",
                        "example": "shopping-cart"
                    },
                    "budget_limit": {
                        "type": "integer",
                        "example": 50000,
                        "nullable": true
                    },
                    "order_index": {
                        "type": "integer"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Authentification via token Sanctum. Utilisez le token reçu lors du login.",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}